bug fixes and cleaning codes
