+++ projects/client/src/Pages/Users/LandingPage.jsx
@@ -54,7 +54,6 @@ const LandingPage = (props) => {
   const fetchProductList = () => {
     let token = localStorage.getItem("tokenIdUser");
     Axios.get(`${API_URL}/users/getproducts`, {
-      headers: { Authorization: `Bearer ${token}` },
     }).then((res) => {
       setProductData(res.data.data);
       console.log(res.data);
