+++ projects/server/src/Routers/userRouters.js
@@ -13,10 +13,10 @@ router.post('/forgot', userControllers.forgot);
 router.patch('/reset', readToken, userControllers.reset);
 router.patch('/edit', readToken, userControllers.edit);
 router.patch('/profilePicture', readToken, userControllers.profilePicture);
-router.get("/getproducts", readToken, userControllers.getProducts);
+router.get("/getproducts", userControllers.getProducts);
 router.get(
   "/getproductdetail/:id",
-  readToken,
+  
   userControllers.getproductDetail
 );
